requirements.txt
pytest
pyyaml
sh